+++ eventstore/api/common/src/main/java/org/occurrent/eventstore/api/WriteResult.java
@@ -18,63 +18,13 @@
 package org.occurrent.eventstore.api;
 
 import java.util.Objects;
-import java.util.StringJoiner;
 
 /**
  * The result of a write operation to the event store.
  */
-public class WriteResult {
+public record WriteResult(String streamId, long oldStreamVersion, long newStreamVersion) {
 
-    private final String streamId;
-    private final long oldStreamVersion;
-    private final long newStreamVersion;
-
-    public WriteResult(String streamId, long oldStreamVersion, long newStreamVersion) {
+    public WriteResult {
         Objects.requireNonNull(streamId, "Stream id cannot be null");
-        this.streamId = streamId;
-        this.oldStreamVersion = oldStreamVersion;
-        this.newStreamVersion = newStreamVersion;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof WriteResult)) return false;
-        WriteResult that = (WriteResult) o;
-        return oldStreamVersion == that.oldStreamVersion && newStreamVersion == that.newStreamVersion && Objects.equals(streamId, that.streamId);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(streamId, oldStreamVersion, newStreamVersion);
-    }
-
-    @Override
-    public String toString() {
-        return new StringJoiner(", ", WriteResult.class.getSimpleName() + "[", "]")
-                .add("streamId='" + streamId + "'")
-                .add("oldStreamVersion=" + oldStreamVersion)
-                .add("newStreamVersion=" + newStreamVersion)
-                .toString();
-    }
-
-    /**
-     * @deprecated Use {@link #getNewStreamVersion()} instead
-     */
-    @Deprecated
-    public long getStreamVersion() {
-        return newStreamVersion;
-    }
-
-    public String getStreamId() {
-        return streamId;
-    }
-
-    public long getOldStreamVersion() {
-        return oldStreamVersion;
-    }
-
-    public long getNewStreamVersion() {
-        return newStreamVersion;
     }
 }