+++ eventstore/mongodb/spring/blocking/src/test/java/org/occurrent/eventstore/mongodb/spring/blocking/SpringMongoEventStoreTest.java
@@ -304,9 +304,9 @@ public class SpringMongoEventStoreTest {
 
             // Then
             assertAll(
-                    () -> assertThat(writeResult.getStreamId()).isEqualTo("name"),
-                    () -> assertThat(writeResult.getOldStreamVersion()).isEqualTo(0L),
-                    () -> assertThat(writeResult.getNewStreamVersion()).isEqualTo(2L)
+                    () -> assertThat(writeResult.streamId()).isEqualTo("name"),
+                    () -> assertThat(writeResult.oldStreamVersion()).isEqualTo(0L),
+                    () -> assertThat(writeResult.newStreamVersion()).isEqualTo(2L)
             );
         }
 
@@ -324,9 +324,9 @@ public class SpringMongoEventStoreTest {
 
             // Then
             assertAll(
-                    () -> assertThat(writeResult.getStreamId()).isEqualTo("name"),
-                    () -> assertThat(writeResult.getOldStreamVersion()).isEqualTo(1L),
-                    () -> assertThat(writeResult.getNewStreamVersion()).isEqualTo(3L)
+                    () -> assertThat(writeResult.streamId()).isEqualTo("name"),
+                    () -> assertThat(writeResult.oldStreamVersion()).isEqualTo(1L),
+                    () -> assertThat(writeResult.newStreamVersion()).isEqualTo(3L)
             );
         }
 
@@ -337,9 +337,9 @@ public class SpringMongoEventStoreTest {
 
             // Then
             assertAll(
-                    () -> assertThat(writeResult.getStreamId()).isEqualTo("name"),
-                    () -> assertThat(writeResult.getOldStreamVersion()).isEqualTo(0L),
-                    () -> assertThat(writeResult.getNewStreamVersion()).isEqualTo(0L)
+                    () -> assertThat(writeResult.streamId()).isEqualTo("name"),
+                    () -> assertThat(writeResult.oldStreamVersion()).isEqualTo(0L),
+                    () -> assertThat(writeResult.newStreamVersion()).isEqualTo(0L)
             );
         }
 
@@ -356,9 +356,9 @@ public class SpringMongoEventStoreTest {
 
             // Then
             assertAll(
-                    () -> assertThat(writeResult.getStreamId()).isEqualTo("name"),
-                    () -> assertThat(writeResult.getOldStreamVersion()).isEqualTo(2L),
-                    () -> assertThat(writeResult.getNewStreamVersion()).isEqualTo(2L)
+                    () -> assertThat(writeResult.streamId()).isEqualTo("name"),
+                    () -> assertThat(writeResult.oldStreamVersion()).isEqualTo(2L),
+                    () -> assertThat(writeResult.newStreamVersion()).isEqualTo(2L)
             );
         }
     }