+++ eventstore/mongodb/spring/reactor/src/test/java/org/occurrent/eventstore/mongodb/spring/reactor/ReactorMongoEventStoreTest.java
@@ -384,9 +384,9 @@ public class ReactorMongoEventStoreTest {
 
             // Then
             assertAll(
-                    () -> assertThat(writeResult.getStreamId()).isEqualTo("name"),
-                    () -> assertThat(writeResult.getOldStreamVersion()).isEqualTo(0L),
-                    () -> assertThat(writeResult.getNewStreamVersion()).isEqualTo(2L)
+                    () -> assertThat(writeResult.streamId()).isEqualTo("name"),
+                    () -> assertThat(writeResult.oldStreamVersion()).isEqualTo(0L),
+                    () -> assertThat(writeResult.newStreamVersion()).isEqualTo(2L)
             );
         }
 
@@ -404,9 +404,9 @@ public class ReactorMongoEventStoreTest {
 
             // Then
             assertAll(
-                    () -> assertThat(writeResult.getStreamId()).isEqualTo("name"),
-                    () -> assertThat(writeResult.getOldStreamVersion()).isEqualTo(1L),
-                    () -> assertThat(writeResult.getNewStreamVersion()).isEqualTo(3L)
+                    () -> assertThat(writeResult.streamId()).isEqualTo("name"),
+                    () -> assertThat(writeResult.oldStreamVersion()).isEqualTo(1L),
+                    () -> assertThat(writeResult.newStreamVersion()).isEqualTo(3L)
             );
         }
 
@@ -417,9 +417,9 @@ public class ReactorMongoEventStoreTest {
 
             // Then
             assertAll(
-                    () -> assertThat(writeResult.getStreamId()).isEqualTo("name"),
-                    () -> assertThat(writeResult.getOldStreamVersion()).isEqualTo(0L),
-                    () -> assertThat(writeResult.getNewStreamVersion()).isEqualTo(0L)
+                    () -> assertThat(writeResult.streamId()).isEqualTo("name"),
+                    () -> assertThat(writeResult.oldStreamVersion()).isEqualTo(0L),
+                    () -> assertThat(writeResult.newStreamVersion()).isEqualTo(0L)
             );
         }
 
@@ -436,9 +436,9 @@ public class ReactorMongoEventStoreTest {
 
             // Then
             assertAll(
-                    () -> assertThat(writeResult.getStreamId()).isEqualTo("name"),
-                    () -> assertThat(writeResult.getOldStreamVersion()).isEqualTo(2L),
-                    () -> assertThat(writeResult.getNewStreamVersion()).isEqualTo(2L)
+                    () -> assertThat(writeResult.streamId()).isEqualTo("name"),
+                    () -> assertThat(writeResult.oldStreamVersion()).isEqualTo(2L),
+                    () -> assertThat(writeResult.newStreamVersion()).isEqualTo(2L)
             );
         }
     }